Improved Player Movement and added most of the features
I improved the player movement to allow for differing speed and more comfortable gameplay. I added the chopping board and made it so the player can finish the combinations and deliver them to the customers and they reward or deduct points as well as time. I also improved the lighting and positioning a little bit.